+++ core/modules/modmassaction.class.php
@@ -91,7 +91,12 @@ class modmassaction extends DolibarrModules
 		//							'workflow' => array('WORKFLOW_MODULE1_YOURACTIONTYPE_MODULE2'=>array('enabled'=>'! empty($conf->module1->enabled) && ! empty($conf->module2->enabled)', 'picto'=>'yourpicto@massaction')) // Set here all workflow context managed by module
 		//                        );
 		$this->module_parts = array(
-			'hooks' => array('data'=>array('main')
+			'hooks' => array(
+//				'data' => array(
+//					'main',
+//				),
+				'propalcard',
+				'globalcard'
 			)
 		);